split config produce to files
